结束下单周期

+++ src/main/java/com/xiliulou/electricity/service/impl/UserServiceImpl.java
@@ -510,6 +510,10 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public R endLimitUser(Long uid) {
+		String orderLimit = redisService.get(ElectricityCabinetConstant.ORDER_TIME_UID+uid);
+		if (Objects.isNull(orderLimit)) {
+			return R.fail("ELECTRICITY.0054", "用户未被限制");
+		}
 		redisService.deleteKeys(ElectricityCabinetConstant.ORDER_TIME_UID +uid);
 		return R.ok();
 	}
